+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - [KOGITO-3215](https://issues.redhat.com/browse/KOGITO-3215) Support cluster-scoped deployment for operator
 - [KOGITO-3937](https://issues.redhat.com/browse/KOGITO-3937) Move probe configuration in to `probes` subproperty
 - [KOGITO-3898](https://issues.redhat.com/browse/KOGITO-3898) Provide CLI flags for subset of probe parameters
+- [KOGITO-3968](https://issues.redhat.com/browse/KOGITO-3968) Add support to build and deploy sw.json and sw.yaml files from source
 
 ## Bug Fixes
 - [KOGITO-3866](https://issues.redhat.com/browse/KOGITO-3866) Operator print error logs when KogitoRuntime delete